+++ games/urls.py
@@ -19,12 +19,6 @@ urlpatterns = [
     path("api/list/<int:game_pk>/playlog/", views.GamePlaytimeAPIView.as_view(), name="playlog"),
     path('api/chatbot/', views.ChatbotAPIView, name='chatbot'),
 
-    path("api/admin/list/", views.AdminGameList, name="game_list"),
-    path("api/list/<int:game_pk>/register/", views.game_register, name="game_register"),
-    path("api/list/<int:game_pk>/deny/", views.game_register_deny, name="game_register_deny"),
-    path("api/denylog/<int:game_pk>/", views.deny_log, name="deny_log"),
-    path('api/list/<int:game_pk>/dzip/', views.game_dzip, name='game_dzip'),
-
     # ---------- Web ---------- #
     # path("list/<int:game_pk>/", views.game_detail_view, name="game_detail_page"),
     # path("list/<int:game_pk>/update/", views.game_update_view, name="game_update_page"),
