+++ games/views.py
@@ -1,19 +1,14 @@
 import re
-import os
-import zipfile
 
-from django.core.files.storage import FileSystemStorage
-from django.http import FileResponse, Http404
-from django.shortcuts import render, get_object_or_404, get_list_or_404, redirect
-from django.db.models import Avg, Q,Count
-from django.db.models.functions import Round
+from django.http import Http404
+from django.shortcuts import get_object_or_404
+from django.db.models import Q, Count
 
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.permissions import IsAuthenticated, AllowAny  # 로그인 인증토큰
 from rest_framework import status
-from rest_framework.pagination import PageNumberPagination
 from rest_framework.decorators import permission_classes
 
 from games.pagination import CategoryGamesPagination, ReviewPagination
@@ -29,7 +24,6 @@ from .models import (
     ReviewsLike,
     PlayLog,
     TotalPlayTime,
-    DenyLog,
 )
 from accounts.models import BotCnt
 from .serializers import (
@@ -43,7 +37,6 @@ from .serializers import (
 from django.conf import settings
 from openai import OpenAI
 from django.utils import timezone
-from datetime import timedelta
 from spartagames.pagination import CustomPagination
 import random
 from urllib.parse import urlencode
@@ -150,6 +143,7 @@ class GameListAPIView(APIView):
         #     data["my_game_pack"] = my_game_pack
         return Response(data, status=status.HTTP_200_OK)
 
+
     """
     게임 등록
     """
@@ -190,7 +184,7 @@ class GameListAPIView(APIView):
             title=request.data.get('title'),
             thumbnail=request.FILES.get('thumbnail'),
             youtube_url=request.data.get('youtube_url'),
-            maker=request.user,  # FE 확인필요
+            maker=request.user,
             content=request.data.get('content'),
             gamefile=request.FILES.get('gamefile'),
             base_control=request.data.get('base_control'),
@@ -218,6 +212,14 @@ class GameListAPIView(APIView):
             )
             screenshots.append(screenshot.src.url)
 
+        # 게임 등록 로그에 데이터 추가
+        game.logs_game.create(
+            recoder = request.user,
+            maker = request.user,
+            game = game,
+            content = "검수요청",
+        )
+        
         return Response({"message": "게임업로드 성공했습니다"}, status=status.HTTP_200_OK)
 
 
@@ -292,6 +294,7 @@ def game_list_search(request):
 
     return paginator.get_paginated_response(response_data)
 
+
 @api_view(['GET'])
 @permission_classes([AllowAny])
 def category_games_list(request):
@@ -329,6 +332,7 @@ def category_games_list(request):
     serializer = GameListSerializer(paginated_games, many=True, context={'user': request.user})
     
     return paginator.get_paginated_response(serializer.data)
+
 
 class GameDetailAPIView(APIView):
     """
@@ -407,7 +411,25 @@ class GameDetailAPIView(APIView):
             if request.data.get('screenshots'):
                 for item in request.FILES.getlist("screenshots"):
                     game.screenshots.create(src=item)
-
+            
+            # 게임 파일 수정인 경우 게임 등록 로그에 데이터 추가
+            if game.register_state == 0:
+                game.logs_game.create(
+                    recoder = request.user,
+                    maker = request.user,
+                    game = game,
+                    content = "수정 후 검수요청",
+                )
+            # 그 외의 항목이 수정됐을 경우 게임 등록 로그에 데이터 추가
+            # 추후 작업 예정
+            # else:
+            #     game.logs_game.create(
+            #         recoder = request.user,
+            #         maker = request.user,
+            #         game = game,
+            #         content = "수정",
+            #     )
+            
             return Response({"messege": "수정이 완료됐습니다"}, status=status.HTTP_200_OK)
         else:
             return Response({"error": "작성자가 아닙니다"}, status=status.HTTP_400_BAD_REQUEST)
@@ -423,6 +445,14 @@ class GameDetailAPIView(APIView):
         if game.maker == request.user or request.user.is_staff == True:
             game.is_visible = False
             game.save()
+            
+            # 게임 삭제 시 게임 등록 로그에 데이터 추가
+            game.logs_game.create(
+                recoder = request.user,
+                maker = request.user,
+                game = game,
+                content = "삭제",
+            )
             return Response({"message": "삭제를 완료했습니다"}, status=status.HTTP_200_OK)
         else:
             return Response({"error": "작성자가 아닙니다"}, status=status.HTTP_400_BAD_REQUEST)
@@ -737,197 +767,10 @@ class GamePlaytimeAPIView(APIView):
             return Response({"error": "게임이 존재하지 않습니다."}, status=status.HTTP_404_NOT_FOUND)
 
 
-@api_view(['POST'])
-def game_register(request, game_pk):
-    # 관리자 여부 확인
-    if request.user.is_staff is False:
-        return Response({"error": "관리자 권한이 필요합니다."}, status=status.HTTP_403_FORBIDDEN)
-
-    # game_pk에 해당하는 row 가져오기 (게시 중인 상태이면서 '등록 중' 상태)
-    row = get_object_or_404(
-        Game, pk=game_pk, is_visible=True, register_state=0)
-
-    # gamefile 필드에 저장한 경로값을 'path' 변수에 저장
-    path = row.gamefile.url
-
-    # ~/<업로드시각>_<압축파일명>.zip 에서 '<업로드시각>_<압축파일명>' 추출
-    game_folder = path.split('/')[-1].split('.')[0]
-
-    # 게임 폴더 경로(압축을 풀 경로): './media/games/<업로드시각>_<압축파일명>'
-    game_folder_path = f"./media/games/{game_folder}"
-
-    # index.html 우선 압축 해제
-    zipfile.ZipFile(f"./{path}").extract("index.html", game_folder_path)
-
-    """
-    index.html 내용 수정
-    <link> 태그 href 값 수정 (line: 7, 8)
-    var buildUrl 변수 값 수정 (line: 59)
-    
-    new_lines: 덮어쓸 내용 저장
-    is_check_build: Build 키워드 찾은 후 True로 변경 (이후 라인에서 Build 찾는 것을 피하기 위함)
-    """
-
-    new_lines = str()
-    is_check_build = False
-
-    # 덮어쓸 내용 담기
-    with open(f"{game_folder_path}/index.html", 'r') as f:
-        for line in f.readlines():
-            if line.find('link') > -1:
-                cursor = line.find('TemplateData')
-                new_lines += line[:cursor] + \
-                    f'/media/games/{game_folder}/' + line[cursor:]
-            elif line.find('buildUrl') > -1 and not is_check_build:
-                is_check_build = True
-                cursor = line.find('Build')
-                new_lines += line[:cursor] + \
-                    f'/media/games/{game_folder}/' + line[cursor:]
-            elif line.find('canvas.style.width') > -1 or line.find('canvas.style.height') > -1:
-                is_check_build = True
-                cursor = line.find('\"')
-                new_lines += line[:cursor] + "\"100%\"\n"
-            else:
-                new_lines += line
-    # 추가할 JavaScript 코드 (iframe의 width, height 조절을 위해 추가함)
-    additional_script = """
-    <script>
-      function sendSizeToParent() {
-        var canvas = document.querySelector("#unity-canvas");
-        var width = canvas.clientWidth;
-        var height = canvas.clientHeight;
-        window.parent.postMessage({ width: width, height: height }, '*');
-      }
-
-      window.addEventListener('resize', sendSizeToParent);
-      window.addEventListener('load', sendSizeToParent);
-    </script>
-    """
-    # CSS 스타일 추가 (body 태그와 unity-container에 overflow: hidden 추가)
-    new_lines = new_lines.replace(
-        '<body', '<body style="margin: 0; padding: 0; width: 100%; height: 100%; overflow: hidden;"')
-    new_lines = new_lines.replace(
-        '<div id="unity-container"', '<div id="unity-container" style="width: 100%; height: 100%; overflow: hidden;"')
-
-    # </body> 태그 전에 추가할 스크립트 삽입
-    body_close_category_index = new_lines.find('</body>')
-    new_lines = new_lines[:body_close_category_index] + \
-        additional_script + new_lines[body_close_category_index:]
-
-    # 덮어쓰기
-    with open(f'{game_folder_path}/index.html', 'w') as f:
-        f.write(new_lines)
-
-    # index.html 외 다른 파일들 압축 해제
-    zipfile.ZipFile(f"./{path}").extractall(
-        path=game_folder_path,
-        members=[item for item in zipfile.ZipFile(
-            f"./{path}").namelist() if item != "index.html"]
-    )
-
-    # 게임 폴더 경로를 저장하고, 등록 상태 1로 변경(등록 성공)
-    row.gamepath = game_folder_path[1:]
-    row.register_state = 1
-    row.save()
-
-    # 알맞은 HTTP Response 리턴
-    # return Response({"message": f"등록을 성공했습니다. (게시물 id: {game_pk})"}, status=status.HTTP_200_OK)
-
-    # 2024-10-31 추가. return 수정 필요 (redirect -> response)
-    # return redirect("games:admin_list")
-    return Response({"message": "게임이 등록되었습니다."}, status=status.HTTP_200_OK)
-
-
-@api_view(['POST'])
-def game_register_deny(request, game_pk):
-    # 관리자 여부 확인
-    if request.user.is_staff is False:
-        return Response({"error": "관리자 권한이 필요합니다."}, status=status.HTTP_403_FORBIDDEN)
-    # # 등록 거부 사유 없을 시 400 (추후 추가)
-    # if request.data.get("deny_content", None) is None:
-    #     return Response({"error": "반려 사유가 필요합니다."}, status=status.HTTP_400_BAD_REQUEST)
-
-    # 게임 정보 수정
-    game = get_object_or_404(
-        Game, pk=game_pk, is_visible=True, register_state=0)
-    game.register_state = 2
-    game.save()
-    
-    # 등록 거부 사유 로그 추가
-    DenyLog.objects.create(
-        admin=request.user,
-        maker=game.maker,
-        game=game,
-        deny_content=request.data.get("deny_content")
-    )
-
-    # 2024-10-31 추가. return 수정 필요 (redirect -> response)
-    # return redirect("games:admin_list")
-    return Response({"message": "게임 등록을 거부했습니다."}, status=status.HTTP_200_OK)
-
-
-@api_view(['POST'])
-def game_dzip(request, game_pk):
-    # 관리자 여부 확인
-    if request.user.is_staff is False:
-        return Response({"error": "관리자 권한이 필요합니다."}, status=status.HTTP_403_FORBIDDEN)
-
-    row = get_object_or_404(
-        Game, pk=game_pk, register_state=0, is_visible=True)
-    zip_path = row.gamefile.url
-    zip_folder_path = "./media/zips/"
-    zip_name = os.path.basename(zip_path)
-
-    # FileSystemStorage 인스턴스 생성
-    # zip_folder_path에 대해 FILE_UPLOAD_PERMISSIONS = 0o644 권한 설정
-    # 파일을 어디서 가져오는지를 정하는 것으로 보면 됨
-    # 디폴트 값: '/media/' 에서 가져옴
-    fs = FileSystemStorage(zip_folder_path)
-
-    # FileResponse 인스턴스 생성
-    # FILE_UPLOAD_PERMISSIONS 권한을 가진 상태로 zip_folder_path 경로 내의 zip_name 파일에 'rb' 모드로 접근
-    # content_type 으로 zip 파일임을 명시
-    response = FileResponse(fs.open(zip_name, 'rb'),
-                            content_type='application/zip')
-
-    # 'Content-Disposition' value 값(HTTP Response 헤더값)을 설정
-    # 파일 이름을 zip_name 으로 다운로드 폴더에 받겠다는 뜻
-    response['Content-Disposition'] = f'attachment; filename="{zip_name}"'
-
-    # FileResponse 객체를 리턴
-    return response
-
-
-# 게임 등록 거부 사유 불러오는 API
-@api_view(['GET'])
-def deny_log(request, game_pk):
-    # 등록 거부된 게임 정보 불러오기
-    game = get_object_or_404(
-        Game, pk=game_pk, is_visible=True, register_state=2)
-    
-    # 관리자, 메이커 여부 확인
-    if not (request.user.is_staff is True or request.user == game.maker):
-        return Response({"error": "관리자 권한이 필요합니다."}, status=status.HTTP_403_FORBIDDEN)
-
-    # 로그 불러오기
-    rows = DenyLog.objects.filter(game=game)
-    if not rows.exists():
-        return Response({"error": "등록 거부된 적이 없는 게임입니다."}, status=status.HTTP_400_BAD_REQUEST)
-    row = rows.order_by("-created_at").first()
-
-    return Response(
-        {
-            "deny_content": row.deny_content
-        },
-        status=status.HTTP_200_OK
-    )
-
-
 CLIENT = OpenAI(api_key=settings.OPEN_API_KEY)
 MAX_USES_PER_DAY = 10  # 하루 당 질문 10개로 제한기준
 
 # chatbot API
-
 
 @api_view(['POST'])
 @permission_classes([IsAuthenticated])
@@ -971,23 +814,6 @@ def ChatbotAPIView(request):
     if about_category in uncategorylist:
         about_category = '없음'
     return Response({"category": about_category}, status=status.HTTP_200_OK)
-
-
-# 관리자용 게임 승인/거절용 리스트
-@api_view(['GET'])
-@permission_classes([IsAuthenticated])
-def AdminGameList(request):
-    if (request.user.is_staff == False) or (request.user.is_superuser == False):
-        return Response({"error": "관리자 권한이 필요합니다."}, status=status.HTTP_403_FORBIDDEN)
-    rows = Game.objects.filter(is_visible=True, register_state=0)
-
-    serializer = GameListSerializer(rows, many=True, context={'user': request.user})
-
-    # 응답 데이터 구성
-    data = {
-        "game_list": serializer.data,
-    }
-    return Response(data, status=status.HTTP_200_OK)
 
 
 # ---------- Web ---------- #
