+++ qnas/models.py
@@ -1,4 +1,7 @@
 from django.db import models
+from django.conf import settings
+
+from games.models import Game
 
 # Create your models here.
 class QnA(models.Model):
@@ -11,3 +14,18 @@ class QnA(models.Model):
     content=models.TextField()
     category = models.CharField(max_length=1, choices=CATEGORY_CHOICES)
     is_visible=models.BooleanField(default=True)
+
+
+# 게임 등록 로그
+class GameRegisterLog(models.Model):
+    recoder = models.ForeignKey(
+        settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name="logs_recoder"
+    )
+    maker = models.ForeignKey(
+        settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name="logs_maker"
+    )
+    game = models.ForeignKey(
+        Game, on_delete=models.CASCADE, related_name="logs_game"
+    )
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
