+++ spartagames/urls.py
@@ -27,7 +27,7 @@ urlpatterns = [
     path('oauth/', include('allauth.urls')),
     path('users/', include('users.urls')),
     path('games/', include('games.urls')),
-    path('qnas/', include('qnas.urls')),
+    path('directs/', include('qnas.urls')),    #2025-01-03 기존 qnas 앱은 놔두고 url 패턴 수정
 
     # ---------- Web ---------- #
     # path('', views.main_view, name='main_view'),
